web/fast-web/fast_sql_al_proj/pyproject.toml
[project]
name = "fast-sql-al-proj"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "psycopg[binary]>=3.3.3",
    "python-decouple>=3.8",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
